Include downloading of XBRL zip data and return FilingMetaData.
Modify _utils.py to for the url and filename for the XBRL zip
file that comes with 10-K, 10-K/A, 10-Q, 10-Q/A filings.

Modify Downloader to download those files in addition to the full
text submission file.

Added the kwarg return_meta to the get function. If it's set
the returned value is a tuple with the FilingMetaData object
provided in addition to the filing count.

Limtations:
Should work with document types mentioned above, but could
break for other document types. Also, XBRL did not start
showing up in filing until about 2011-(ish) so it would
make sense to place guards on document type and or make
the XBRL detail files an option. Open to suggestion.

Relates to #32 because the FilingMetaData object is returned,
optionally. Believe this is the best option because the files
can be downloaded and the data returned in one web request
instead of two.