Update _orchestrator.py
Hi, 
I want to suggest a change to make the text files easier to get to for further manipulation. I'm trying to shorten the Path and mark the downloaded file directly with the company identifier and accession number. 
I'm somewhat new to this, and this is my first attempt at creating a fork. Please let me know if I missed anything. 
Thanks, 
TAB